test-sbt: match tests on both short and prefixed names. (#9756)
If a test is defined as `suite("suite")(test("test"){})`
it seems reasonable that we should be able to ask for it to be executed
by either its short name "test" or by its full name "suite - test".
This *can* be achieved by stripping the suite prefix "suite -" from the names from the selectors.
However, this approach breaks down in the presence of nested suites;
e.g. if a test is defined as `suite("outer")(suite("inner)(test("test"){}))`,
when the outer suite is run:
- selector "test" still executes the test - as it should;
- selector "inner - test" does not execute the test (wrong prefix gets stipped) - but should;
- selector "outer - inner - test" does not execute the test (inner prefix does not get stripped) - but should;
- selector "outer - test" does execute the test - but shouldn't, since test with such a full name does not exist.

 To make this work correctly, `Spec.filterLabels()` has to be adjusted to keep track of the accumulated
suite prefix and match the search terms against the full names.
Once that is done, we do not need to strip anything here;
in fact, we do not even need know what the label of the spec is.